Switching the conversation view controller, to just use a conversation controller.
