+++ iOS/Flows/Conversation/Conversation List/ConversationViewController+Extensions.swift
@@ -12,22 +12,6 @@ import StreamChat
 extension ConversationViewController {
 
     func setupInputHandlers() {
-        self.dataSource.handleDidTapClose = { [unowned self] item in
-            self.dataSource.deleteItems([item])
-            switch item {
-            case .upsell:
-                UserDefaultsManager.update(key: .shouldShowGroupsUpsell, with: false)
-            case .invest:
-                UserDefaultsManager.update(key: .shouldShowInvestUpsell, with: false)
-            default:
-                break
-            }
-        }
-
-        self.dataSource.handleLoadMoreMessages = { [unowned self] in
-            self.loadMoreConversationsIfNeeded()
-        }
-        
         self.dataSource.handleCollectionViewTapped = { [unowned self] in
             if self.messageInputController.swipeInputView.textView.isFirstResponder {
                 self.messageInputController.swipeInputView.textView.resignFirstResponder()
@@ -69,13 +53,14 @@ extension ConversationViewController {
     }
     
     func subscribeToConversationUpdates() {
-        self.conversationListController
-            .channelsChangesPublisher
-            .mainSink { [unowned self] _ in
-                Task {
-                    await self.dataSource.update(with: self.conversationListController)
-                }.add(to: self.autocancelTaskPool)
-            }.store(in: &self.cancellables)
+        #warning("restore")
+//        self.conversationListController
+//            .channelsChangesPublisher
+//            .mainSink { [unowned self] _ in
+//                Task {
+//                    await self.dataSource.update(with: self.conversationListController)
+//                }.add(to: self.autocancelTaskPool)
+//            }.store(in: &self.cancellables)
 
         self.messageInputController.swipeInputView.textView.$inputText.mainSink { [unowned self] text in
             guard let conversationController = self.getCurrentConversationController() else { return }
